refactor(runtime): extract shell_exec tool module
tool_shell_exec moves into tool_runner/shell.rs verbatim. All security
gating (taint sinks via tool_runner::taint, RO-workspace verb
classification via tool_runner::shell_safety) stays upstream in the
dispatcher — by the time we reach this function the command has
already been cleared, so the new module pulls in no super::* helper
beyond the standard `crate::*` references it already had.

Both execution paths (Allowlist direct-exec, Full sh -c) keep their
original byte-for-byte logic, including the pipe-draining
wait_with_output() + 100ms interrupt-tick loop that #4759 introduced
to fix the >8KB-output deadlock and preserve cancel-cascade behavior.

mod.rs: 10,889 -> 10,699 LOC (-190). cargo check + cargo clippy --lib
both clean.

Refs #3710
